+++ storeNodes.java
@@ -1,14 +1,14 @@
 import java.util.*;
 
 public class storeNodes {
-	private List<Integer> arr;
-	public storeNodes(int n){
-		arr = new ArrayList<Integer>();
+	private List<Double> arr;
+	public storeNodes(double n){
+		arr = new ArrayList<Double>();
 	}
-	public void insert(int t){
+	public void insert(double t){
 		arr.add(t);
 	}
-	public boolean find(int x){
+	public boolean find(double x){
 		return arr.contains(x);
 	}
 }
